% Generated by roxygen2 (4.0.2): do not edit by hand
\name{k.capital}
\alias{k.capital}
\alias{rwa}
\title{Risk-weighted Asset (RWA) and the required capital, k, calculation for retail portfolios under the
BasellII framework}
\usage{
k.capital(pd, lgd, asset.class = c("Mortgage", "QRRE", "Retail Other"),
  r = NULL)

rwa(..., ead = 1, reg.mult = 1)
}
\arguments{
\item{pd}{The probability of default (PD) of the record}

\item{lgd}{The loss given default (LGD) of the record}

\item{asset.class}{The assest class of the account}

\item{r}{The assest correlation. Usually not required as it is specifeid under BaselII bassed on asset class}

\item{ead}{The exposure at default in dollar amount. If not specified the
function returns the RWA\%}

\item{reg.mult}{Regulator specified multiplier to the RWA}
}
\description{
Risk-weighted Asset (RWA) and the required capital, k, calculation for retail portfolios under the
BasellII framework
}
\examples{
rwa(pd = 0.01, lgd = 0.2)

# sometimes the regulator may impose a different correlation to the Basel II
# standard. YOu can use r to get the correaltion
rwa(pd = 0.01, lgd = 0.2, r = 0.21)
}

