% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rwa.R
\name{rwa}
\alias{asset.corr}
\alias{k.capital}
\alias{rwa}
\title{Risk-weighted Asset (RWA) and the required capital, k, calculation for
  retail portfolios under the Basel II framework}
\usage{
rwa(..., ead = 1, reg.mult = 1)

asset.corr(asset.class = c("Mortgage", "QRRE", "Other Retail"), pd = NULL)

k.capital(pd, lgd, asset.class = c("Mortgage", "QRRE", "Other Retail"),
  r = NULL)
}
\arguments{
\item{...}{arguments that feed into the k.capital function}

\item{ead}{The exposure at default in dollar amount. If not specified the
function returns the RWA\%}

\item{reg.mult}{Regulator specified multiplier to the RWA}

\item{asset.class}{The assest class.}

\item{pd}{The probability of default (PD) of the record}

\item{lgd}{The loss given default (LGD) of the record}

\item{r}{The assest correlation. Usually not required as it is specifeid under Basel II bassed on asset class}
}
\value{
The asset correlation correpoding to the asset class as specified in
   Basel II

The capital requirement as specifed under Basel II A-IRB
}
\description{
Risk-weighted Asset (RWA) and the required capital, k, calculation for
  retail portfolios under the Basel II framework
}
\examples{
rwa(pd = 0.01, lgd = 0.2)
}

