% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gini.R
\name{gini}
\alias{gini}
\title{GINI coefficient for binary targets}
\usage{
gini(score, target, bad = NULL, return.data = T, return.data.table = F)
}
\arguments{
\item{score}{A vector of numbers that is used to predict the target}

\item{target}{A binary vector of targets. The first (alphanumerically) sorted level is used as the good.}

\item{bad}{The bad target. It must be one of the two values in target}

\item{return.data}{Return the data and the computation}

\item{return.data.table}{Return the data as data.table. By default this is false because data.frame is more common.}
}
\description{
GINI coefficient for binary targets
}

