% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psi.R
\name{psi}
\alias{psi}
\title{Population Stability Index (PSI)}
\usage{
psi(base.segments, comparison.segments)
}
\arguments{
\item{base.segments}{The base segment's distribution for comparison. It needs
to be a vector of positive numbers in the open interval (0,1) that sums to
exactly 1.}

\item{comparison.segments}{The distribution of the segments that is to be
compared to the base.segments. It needs to be a vector of positive numbers
in the open interval (0,1) that sums to exactly 1 and also be of the same
length as base.segments}
}
\description{
Population Stability Index (PSI)
}
\examples{
x <- runif(10)
base.seg <- x/sum(x)
# this should return a PSI of zero since the segments' distributions did not change
psi(base.seg, base.seg)
}

